+++ GrowthBook/src/jvmTest/kotlin/com/sdk/growthbook/integration/VerifySDKReturnFeatureValues.kt
@@ -25,7 +25,7 @@ import com.sdk.growthbook.kotlinx.serialization.gbSerialize
 internal class VerifySDKReturnFeatureValues {
 
     @Test
-    fun verifySDKReturnFeatureDefaultValue() {
+    fun verifySDKReturnFeatureDefaultValue() = runTest {
         @Language("json")
         val json = """
             {
@@ -63,7 +63,7 @@ internal class VerifySDKReturnFeatureValues {
     }
 
     @Test
-    fun verifySDKReturnFeatureValueByConditionIfAttributeDoesNotExist() {
+    fun verifySDKReturnFeatureValueByConditionIfAttributeDoesNotExist() = runTest {
         @Language("json")
         val json = """
             {
@@ -102,7 +102,7 @@ internal class VerifySDKReturnFeatureValues {
     }
 
     @Test
-    fun verifySDKAttributesCastingTypes() {
+    fun verifySDKAttributesCastingTypes() = runTest {
         @Language("json")
         val json = """
 {
@@ -159,7 +159,7 @@ internal class VerifySDKReturnFeatureValues {
     }
 
     @Test
-    fun `if features fetch fails, suspendFeature() method calls fetchFeature() again`() {
+    fun `if features fetch fails, suspendFeature() method calls fetchFeature() again`() = runTest {
         val gbSdk = buildSDK(json = "{}", attributes = emptyMap()) // buildSDK() calls refreshCache(),
         // refreshCache() calls fetchFeature()
         // fetchFeature() triggers featuresFetchFailed()
@@ -169,21 +169,19 @@ internal class VerifySDKReturnFeatureValues {
         }
         gbSdk.featuresViewModel = mockedFeaturesViewModel
 
-        runTest {
-            val job = launch {
-                gbSdk.suspendFeature("some-feature-id")
-            }
-
-            // it is not mandatory here but just to emphasize that
-            // if call failed, then suspendFeature() method calls fetchFeatures()
-            gbSdk.featuresFetchFailed(GBError(null), true)
-
-            delay(100) // cancel only after 100 millis of waiting
-            job.cancel()
-            // or gbSdk.featuresFetchedSuccessfully(emptyMap(), true)
-
-            coVerify { mockedFeaturesViewModel.fetchFeatures() }
+        val job = launch {
+            gbSdk.suspendFeature("some-feature-id")
         }
+
+        // it is not mandatory here but just to emphasize that
+        // if call failed, then suspendFeature() method calls fetchFeatures()
+        gbSdk.featuresFetchFailed(GBError(null), true)
+
+        delay(100) // cancel only after 100 millis of waiting
+        job.cancel()
+        // or gbSdk.featuresFetchedSuccessfully(emptyMap(), true)
+
+        coVerify { mockedFeaturesViewModel.fetchFeatures() }
     }
 
 /*