+++ GrowthBook/src/jvmTest/kotlin/com/sdk/growthbook/tests/FeaturesViewModelTests.kt
@@ -10,9 +10,14 @@ import com.sdk.growthbook.features.FeaturesViewModel
 import com.sdk.growthbook.model.GBContext
 import com.sdk.growthbook.model.GBNumber
 import com.sdk.growthbook.model.GBOptions
-import kotlinx.coroutines.Dispatchers
+import com.sdk.growthbook.model.GBString
+import com.sdk.growthbook.stickybucket.GBStickyBucketService
+import com.sdk.growthbook.utils.GBStickyAssignmentsDocument
+import com.sdk.growthbook.utils.GBUtils
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.test.StandardTestDispatcher
+import kotlinx.coroutines.test.UnconfinedTestDispatcher
+import kotlinx.coroutines.test.TestScope
 import kotlinx.coroutines.test.advanceUntilIdle
 import kotlinx.coroutines.test.runCurrent
 import kotlinx.coroutines.test.runTest
@@ -45,18 +50,18 @@ class FeaturesViewModelTests : FeaturesFlowDelegate {
     private val testGbOptions = GBOptions("https://example.com", null)
 
     @Test
-    fun testSuccess() {
+    fun testSuccess() = runTest {
         isSuccess = false
         isError = true
         val viewModel = FeaturesViewModel(
-            this,
+            this@FeaturesViewModelTests,
             FeaturesDataSource(
                 MockNetworkClient(MockResponse.successResponse, null),
                 gbContext,
                 testGbOptions
             ),
             "3tfeoyW0wlo47bDnbWDkxg==", false,
-            coroutineContext = Dispatchers.Unconfined,
+            coroutineContext = UnconfinedTestDispatcher(testScheduler),
         )
 
         viewModel.fetchFeatures()
@@ -67,11 +72,11 @@ class FeaturesViewModelTests : FeaturesFlowDelegate {
     }
 
     @Test
-    fun testSuccessForEncryptedFeatures() {
+    fun testSuccessForEncryptedFeatures() = runTest {
         isSuccess = false
         isError = true
         val viewModel = FeaturesViewModel(
-            this,
+            this@FeaturesViewModelTests,
             FeaturesDataSource(
                 MockNetworkClient(
                     MockResponse.successResponseEncryptedFeatures, null
@@ -79,7 +84,7 @@ class FeaturesViewModelTests : FeaturesFlowDelegate {
                 gbContext, testGbOptions,
             ),
             "3tfeoyW0wlo47bDnbWDkxg==", false,
-            coroutineContext = Dispatchers.Unconfined,
+            coroutineContext = UnconfinedTestDispatcher(testScheduler),
         )
 
         viewModel.fetchFeatures()
@@ -90,12 +95,12 @@ class FeaturesViewModelTests : FeaturesFlowDelegate {
     }
 
     @Test
-    fun testError() {
+    fun testError() = runTest {
 
         isSuccess = false
         isError = true
         val viewModel = FeaturesViewModel(
-            delegate = this,
+            delegate = this@FeaturesViewModelTests,
             dataSource = FeaturesDataSource(
                 MockNetworkClient(
                     null, Throwable("UNKNOWN", null)
@@ -103,7 +108,7 @@ class FeaturesViewModelTests : FeaturesFlowDelegate {
                 gbContext, testGbOptions,
             ),
             cachingEnabled = false,
-            coroutineContext = Dispatchers.Unconfined,
+            coroutineContext = UnconfinedTestDispatcher(testScheduler),
         )
 
         viewModel.fetchFeatures()
@@ -114,12 +119,12 @@ class FeaturesViewModelTests : FeaturesFlowDelegate {
     }
 
     @Test
-    fun testInvalid() {
+    fun testInvalid() = runTest {
 
         isSuccess = false
         isError = true
         val viewModel = FeaturesViewModel(
-            delegate = this,
+            delegate = this@FeaturesViewModelTests,
             dataSource = FeaturesDataSource(
                 MockNetworkClient(
                     MockResponse.ERROR_RESPONSE, null
@@ -128,7 +133,7 @@ class FeaturesViewModelTests : FeaturesFlowDelegate {
             ),
             encryptionKey = "",
             cachingEnabled = false,
-            coroutineContext = Dispatchers.Unconfined,
+            coroutineContext = UnconfinedTestDispatcher(testScheduler),
         )
         viewModel.fetchFeatures()
 
@@ -138,12 +143,12 @@ class FeaturesViewModelTests : FeaturesFlowDelegate {
     }
 
     @Test
-    fun testForRemoteEvalSuccess() {
+    fun testForRemoteEvalSuccess() = runTest {
         isSuccess = false
         isError = true
 
         val viewModel = FeaturesViewModel(
-            delegate = this,
+            delegate = this@FeaturesViewModelTests,
             dataSource =
                 FeaturesDataSource(
                     dispatcher = MockNetworkClient(
@@ -154,7 +159,7 @@ class FeaturesViewModelTests : FeaturesFlowDelegate {
                 ),
             encryptionKey = "3tfeoyW0wlo47bDnbWDkxg==",
             cachingEnabled = false,
-            coroutineContext = Dispatchers.Unconfined,
+            coroutineContext = UnconfinedTestDispatcher(testScheduler),
         )
         val forcedFeature = mapOf("feature" to GBNumber(123))
         val forcedVariation = mapOf("feature" to 123)
@@ -172,12 +177,12 @@ class FeaturesViewModelTests : FeaturesFlowDelegate {
     }
 
     @Test
-    fun testForRemoteEvalFailed() {
+    fun testForRemoteEvalFailed() = runTest {
         isSuccess = false
         isError = true
 
         val viewModel = FeaturesViewModel(
-            delegate = this,
+            delegate = this@FeaturesViewModelTests,
             dataSource =
                 FeaturesDataSource(
                     dispatcher = MockNetworkClient(
@@ -188,7 +193,7 @@ class FeaturesViewModelTests : FeaturesFlowDelegate {
                 ),
             encryptionKey = "3tfeoyW0wlo47bDnbWDkxg==",
             cachingEnabled = false,
-            coroutineContext = Dispatchers.Unconfined,
+            coroutineContext = UnconfinedTestDispatcher(testScheduler),
         )
         val forcedFeature = mapOf("feature" to GBNumber(123))
         val forcedVariation = mapOf("feature" to 123)
@@ -207,19 +212,19 @@ class FeaturesViewModelTests : FeaturesFlowDelegate {
     }
 
     @Test
-    fun testNotModified() {
+    fun testNotModified() = runTest {
         isSuccess = false
         isError = false
         isNotModified = false
 
         val viewModel = FeaturesViewModel(
-            delegate = this,
+            delegate = this@FeaturesViewModelTests,
             dataSource = FeaturesDataSource(
                 MockNetworkClient(successResponse = null, error = null, notModified = true),
                 gbContext, testGbOptions,
             ),
             cachingEnabled = false,
-            coroutineContext = Dispatchers.Unconfined,
+            coroutineContext = UnconfinedTestDispatcher(testScheduler),
         )
 
         viewModel.fetchFeatures()
@@ -230,11 +235,11 @@ class FeaturesViewModelTests : FeaturesFlowDelegate {
     }
 
     @Test
-    fun testSuccessWithCachingEnabled() {
+    fun testSuccessWithCachingEnabled() = runTest {
         isSuccess = false
         isError = true
         val viewModel = FeaturesViewModel(
-            this,
+            this@FeaturesViewModelTests,
             FeaturesDataSource(
                 MockNetworkClient(MockResponse.successResponse, null),
                 gbContext,
@@ -242,7 +247,7 @@ class FeaturesViewModelTests : FeaturesFlowDelegate {
             ),
             "3tfeoyW0wlo47bDnbWDkxg==",
             cachingEnabled = true,
-            coroutineContext = Dispatchers.Unconfined,
+            coroutineContext = UnconfinedTestDispatcher(testScheduler),
         )
 
         viewModel.fetchFeatures()
@@ -253,10 +258,10 @@ class FeaturesViewModelTests : FeaturesFlowDelegate {
     }
 
     @Test
-    fun testFeaturesAPIModelSuccessfullyCalled() {
+    fun testFeaturesAPIModelSuccessfullyCalled() = runTest {
         featuresAPIModelCalled = false
         val viewModel = FeaturesViewModel(
-            this,
+            this@FeaturesViewModelTests,
             FeaturesDataSource(
                 MockNetworkClient(MockResponse.successResponse, null),
                 gbContext,
@@ -264,7 +269,7 @@ class FeaturesViewModelTests : FeaturesFlowDelegate {
             ),
             "3tfeoyW0wlo47bDnbWDkxg==",
             cachingEnabled = false,
-            coroutineContext = Dispatchers.Unconfined,
+            coroutineContext = UnconfinedTestDispatcher(testScheduler),
         )
 
         viewModel.fetchFeatures()
@@ -273,19 +278,19 @@ class FeaturesViewModelTests : FeaturesFlowDelegate {
     }
 
     @Test
-    fun testSavedGroupsFetchedSuccessfully() {
+    fun testSavedGroupsFetchedSuccessfully() = runTest {
         isSuccess = false
         isError = true
 
         val viewModel = FeaturesViewModel(
-            delegate = this,
+            delegate = this@FeaturesViewModelTests,
             dataSource = FeaturesDataSource(
                 MockNetworkClient(MockResponse.successResponseWithSavedGroups, null),
                 gbContext, testGbOptions,
             ),
             encryptionKey = "",
             cachingEnabled = false,
-            coroutineContext = Dispatchers.Unconfined,
+            coroutineContext = UnconfinedTestDispatcher(testScheduler),
         )
 
         viewModel.fetchFeatures()
@@ -295,18 +300,18 @@ class FeaturesViewModelTests : FeaturesFlowDelegate {
     }
 
     @Test
-    fun testSavedGroupsFetchFailed() {
+    fun testSavedGroupsFetchFailed() = runTest {
         isSuccess = false
         isError = true
         val viewModel = FeaturesViewModel(
-            delegate = this,
+            delegate = this@FeaturesViewModelTests,
             dataSource = FeaturesDataSource(
                 MockNetworkClient(MockResponse.successResponseWithEncryptedFeaturesOnly, null),
                 gbContext, testGbOptions,
             ),
             encryptionKey = "",
             cachingEnabled = false,
-            coroutineContext = Dispatchers.Unconfined,
+            coroutineContext = UnconfinedTestDispatcher(testScheduler),
         )
 
         viewModel.fetchFeatures()
@@ -316,11 +321,11 @@ class FeaturesViewModelTests : FeaturesFlowDelegate {
     }
 
     @Test
-    fun testHandleFetchFeaturesWithoutRemoteEvalPlainFeatures() {
+    fun testHandleFetchFeaturesWithoutRemoteEvalPlainFeatures() = runTest {
         receivedFromCache = false
         val cacheLayer = MockCachingLayer.fromApiResponse(MockResponse.successResponse)
         val viewModel = FeaturesViewModel(
-            delegate = this,
+            delegate = this@FeaturesViewModelTests,
             dataSource = FeaturesDataSource(
                 MockNetworkClient(MockResponse.successResponse, null),
                 gbContext, testGbOptions,
@@ -328,7 +333,7 @@ class FeaturesViewModelTests : FeaturesFlowDelegate {
             encryptionKey = "3tfeoyW0wlo47bDnbWDkxg==",
             cachingEnabled = false,
             cachingLayer = cacheLayer,
-            coroutineContext = Dispatchers.Unconfined,
+            coroutineContext = UnconfinedTestDispatcher(testScheduler),
         )
 
         viewModel.fetchFeatures()
@@ -341,12 +346,12 @@ class FeaturesViewModelTests : FeaturesFlowDelegate {
     }
 
     @Test
-    fun testHandleFetchFeaturesWithoutRemoteEvalEncryptedFeatures() {
+    fun testHandleFetchFeaturesWithoutRemoteEvalEncryptedFeatures() = runTest {
         receivedFromCache = false
         val cacheLayer =
             MockCachingLayer.fromApiResponse(MockResponse.successResponseEncryptedFeatures)
         val viewModel = FeaturesViewModel(
-            delegate = this,
+            delegate = this@FeaturesViewModelTests,
             dataSource = FeaturesDataSource(
                 MockNetworkClient(MockResponse.successResponseEncryptedFeatures, null),
                 gbContext, testGbOptions,
@@ -354,7 +359,7 @@ class FeaturesViewModelTests : FeaturesFlowDelegate {
             encryptionKey = "3tfeoyW0wlo47bDnbWDkxg==",
             cachingEnabled = false,
             cachingLayer = cacheLayer,
-            coroutineContext = Dispatchers.Unconfined,
+            coroutineContext = UnconfinedTestDispatcher(testScheduler),
         )
 
         viewModel.fetchFeatures()
@@ -367,18 +372,18 @@ class FeaturesViewModelTests : FeaturesFlowDelegate {
     }
 
     @Test
-    fun testFetchFeaturesWithCacheException() {
+    fun testFetchFeaturesWithCacheException() = runTest {
         receivedCacheError = false
         val cacheLayer = MockCachingLayer(throwOnGet = true)
         val viewModel = FeaturesViewModel(
-            delegate = this,
+            delegate = this@FeaturesViewModelTests,
             dataSource = FeaturesDataSource(
                 MockNetworkClient(null, Throwable("Network error")),
                 gbContext, testGbOptions,
             ),
             cachingEnabled = false,
             cachingLayer = cacheLayer,
-            coroutineContext = Dispatchers.Unconfined,
+            coroutineContext = UnconfinedTestDispatcher(testScheduler),
         )
 
         viewModel.fetchFeatures()
@@ -392,12 +397,12 @@ class FeaturesViewModelTests : FeaturesFlowDelegate {
     }
 
     @Test
-    fun testCacheWriteFailureDoesNotDiscardFetchedFeatures() {
+    fun testCacheWriteFailureDoesNotDiscardFetchedFeatures() = runTest {
         isSuccess = false
         isError = false
         val cacheLayer = MockCachingLayer(throwOnPut = true)
         val viewModel = FeaturesViewModel(
-            delegate = this,
+            delegate = this@FeaturesViewModelTests,
             dataSource = FeaturesDataSource(
                 MockNetworkClient(MockResponse.successResponse, null),
                 gbContext, testGbOptions,
@@ -405,6 +410,7 @@ class FeaturesViewModelTests : FeaturesFlowDelegate {
             encryptionKey = "3tfeoyW0wlo47bDnbWDkxg==",
             cachingEnabled = true,
             cachingLayer = cacheLayer,
+            coroutineContext = UnconfinedTestDispatcher(testScheduler)
         )
 
         viewModel.fetchFeatures()
@@ -415,16 +421,16 @@ class FeaturesViewModelTests : FeaturesFlowDelegate {
     }
 
     @Test
-    fun testAutoRefreshFeaturesReturnsFlow() {
+    fun testAutoRefreshFeaturesReturnsFlow() = runTest {
         val viewModel = FeaturesViewModel(
-            delegate = this,
+            delegate = this@FeaturesViewModelTests,
             dataSource = FeaturesDataSource(
                 MockNetworkClient(MockResponse.successResponse, null),
                 gbContext, testGbOptions,
             ),
             encryptionKey = "3tfeoyW0wlo47bDnbWDkxg==",
             cachingEnabled = false,
-            coroutineContext = Dispatchers.Unconfined,
+            coroutineContext = UnconfinedTestDispatcher(testScheduler),
         )
 
         val flow = viewModel.autoRefreshFeatures()
@@ -435,7 +441,7 @@ class FeaturesViewModelTests : FeaturesFlowDelegate {
     /**
      * Proves the race-condition fix: features are applied to context only AFTER the sticky-bucket
      * refresh (onPayloadReady) has fully completed — verified under a NON-immediate dispatcher so
-     * the ordering is enforced by code, not masked by Dispatchers.Unconfined running everything inline.
+     * the ordering is enforced by code, not masked by UnconfinedTestDispatcher(testScheduler) running everything inline.
      */
     @OptIn(kotlinx.coroutines.ExperimentalCoroutinesApi::class)
     @Test
@@ -499,6 +505,105 @@ class FeaturesViewModelTests : FeaturesFlowDelegate {
             listOf("stickyRefreshStart", "stickyRefreshDone", "featuresApplied"),
             events,
             "featuresFetchedSuccessfully must fire strictly after onPayloadReady completes"
+        )
+    }
+
+    /**
+     * Integration-level companion to [testFeaturesAppliedOnlyAfterStickyRefreshCompletes]: instead of a
+     * fake onPayloadReady that merely delays, this drives the REAL [GBUtils.refreshStickyBuckets] over the
+     * actual payload (mirroring GrowthBookSDK.onPayloadReady) and asserts that, by the time features are
+     * applied, the user's sticky-bucket docs are already written into the context. Only the storage leaf
+     * (getAllAssignments) is faked — and it suspends on the test scheduler, so the ordering is enforced by
+     * the production code under a NON-immediate dispatcher, not masked by UnconfinedTestDispatcher(testScheduler).
+     */
+    @OptIn(kotlinx.coroutines.ExperimentalCoroutinesApi::class)
+    @Test
+    fun testStickyDocsPopulatedBeforeFeaturesAppliedWithRealRefresh() = runTest {
+        // Fake only the storage IO. It suspends (delay) on the test scheduler to mimic a read that does
+        // not complete synchronously, and returns a doc for whichever identifier attributes the real
+        // refresh derived from the payload.
+        val service = object : GBStickyBucketService {
+            override val coroutineScope = TestScope(testScheduler)
+            override suspend fun getAssignments(
+                attributeName: String,
+                attributeValue: String
+            ): GBStickyAssignmentsDocument? = null
+
+            override suspend fun saveAssignments(doc: GBStickyAssignmentsDocument) = Unit
+
+            override suspend fun getAllAssignments(
+                attributes: Map<String, String>
+            ): Map<String, GBStickyAssignmentsDocument> {
+                delay(100) // storage read in flight — refresh has NOT completed yet
+                return attributes.entries.associate { (name, value) ->
+                    "$name||$value" to GBStickyAssignmentsDocument(
+                        attributeName = name,
+                        attributeValue = value,
+                        assignments = mapOf("exp__0" to "control"),
+                    )
+                }
+            }
+        }
+
+        val ctx = GBContext(
+            apiKey = "Key",
+            enabled = true,
+            attributes = mapOf("id" to GBString("u1")),
+            forcedVariations = emptyMap(),
+            qaMode = false,
+            trackingCallback = { _, _ -> },
+            encryptionKey = null,
+            stickyBucketService = service,
+        )
+
+        // Snapshot of the context's sticky docs at the exact moment features are applied.
+        var docsAtApplyTime: Map<String, GBStickyAssignmentsDocument>? = null
+
+        val delegate = object : FeaturesFlowDelegate {
+            // Mirrors GrowthBookSDK.onPayloadReady: real refresh against the freshly fetched payload.
+            override suspend fun onPayloadReady(model: FeaturesDataModel) {
+                GBUtils.refreshStickyBuckets(
+                    context = ctx,
+                    data = model,
+                    attributeOverrides = emptyMap(),
+                )
+            }
+
+            override fun featuresFetchedSuccessfully(features: GBFeatures, isRemote: Boolean) {
+                docsAtApplyTime = ctx.stickyBucketAssignmentDocs
+            }
+
+            override fun featuresFetchFailed(error: GBError, isRemote: Boolean) = Unit
+            override fun savedGroupsFetchFailed(error: GBError, isRemote: Boolean) = Unit
+            override fun savedGroupsFetchedSuccessfully(savedGroups: JsonObject, isRemote: Boolean) = Unit
+            override fun featuresNotModified() = Unit
+        }
+
+        val viewModel = FeaturesViewModel(
+            delegate = delegate,
+            dataSource = FeaturesDataSource(
+                MockNetworkClient(MockResponse.successResponse, null),
+                ctx, testGbOptions,
+            ),
+            cachingEnabled = false,
+            coroutineContext = StandardTestDispatcher(testScheduler),
+        )
+
+        viewModel.fetchFeatures()
+
+        // Queued, not run: features have not been applied while the refresh is still pending.
+        assertEquals(null, docsAtApplyTime)
+
+        advanceUntilIdle()
+
+        // featuresFetchedSuccessfully fired only after the real refresh wrote the current user's docs.
+        assertNotNull(
+            docsAtApplyTime,
+            "stickyBucketAssignmentDocs must be populated before features are applied",
+        )
+        assertTrue(
+            docsAtApplyTime!!.containsKey("id||u1"),
+            "Refresh must have loaded the current user's assignment doc before features applied",
         )
     }